refactor: side-menu 이름 변경 및 index 파일로 교체
