+++ src/app/test/stateBadge/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from 'react';
 import CardList from '@/components/cardList/CardList';
-import SideMenu from '@/components/sideMenu/SideMenu';
+import SideMenu from '@/components/side-menu';
 import StateBadge from '@/components/stateBadge/StateBadge';
 import Input from '@/components/ui/Input/Input';
 
@@ -65,14 +65,18 @@ export default function StateBadgeTestPage() {
             type='email'
             value={emailValue}
             error={emailError}
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) => { setEmailValue(e.target.value); }}
+            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+              setEmailValue(e.target.value);
+            }}
           />
           <Input
             label='비밀번호'
             placeholder='비밀번호를 입력해 주세요'
             type='password'
             value={passwordValue}
-            onChange={(e: React.ChangeEvent<HTMLInputElement>) => { setEmailValue(e.target.value); }}
+            onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+              setEmailValue(e.target.value);
+            }}
           />
         </div>
       </div>
