+++ src/components/side-menu/validateImageFile.ts
